+++ apps/backend/src/main/java/io/github/artsobol/kurkod/web/domain/breed/model/dto/BreedDTO.java
@@ -1,5 +1,5 @@
 package io.github.artsobol.kurkod.web.domain.breed.model.dto;
 
 public record BreedDTO(
-        Integer id, String name, Integer eggsNumber, Integer weight, Long version
+        Long id, String name, Integer eggsNumber, Integer weight, Long version
 ) {};
